Move `Cycle` to dedicated module
